Moved Mvar and VarU into separate mods
